apply some clang-tidy suggestions

+++ src/Engine/Sound.h
@@ -100,7 +100,7 @@ private:
 	void *m_mixChunkOrMixMusic;
 	unsigned long m_iPrevPosition;
 #if defined(MCENGINE_FEATURE_SOUND)
-    inline long unsigned int _BASS_FX_TempoCreate(long unsigned int chan, long unsigned int flags) const {return DL_CALL_FCT((SoundEngine::m_BASS_FX_TempoCreate), (chan, flags));}
+    inline uint32_t _BASS_FX_TempoCreate(uint32_t chan, uint32_t flags) const {return DL_CALL_FCT((SoundEngine::m_BASS_FX_TempoCreate), (chan, flags));}
 #endif
 };
